pyproject.toml
[project]
name = "spotjax"
version = "0.1.0"
description = "CLI tool for orchestrating JAX training on Google Cloud Spot TPUs"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "SpotJAX Team" }]
keywords = ["jax", "tpu", "gcp", "spot", "training", "ml"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "asyncssh>=2.14.0",
    "google-cloud-tpu>=1.18.0",
    "google-cloud-compute>=1.14.0",
    "google-cloud-storage>=2.14.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
    "jax>=0.6.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]

[project.scripts]
spotjax = "spotjax.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/spotjax"]

[tool.hatch.build.targets.wheel.force-include]
"src/spotjax/templates" = "spotjax/templates"

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["spotjax"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
